Update formatter tests for ktfmt
